+++ CHANGES.rst
@@ -1,10 +1,22 @@
 Changelog
 =========
 
-0.12 (unreleased)
+
+0.12.0 (unreleased)
 -------------------
 
+- Add ``TChannel.is_listening()`` to determine if ``listen`` has been called.
+- Calling ``TChannel.listen()`` more than once raises a ``tchannel.errors.AlreadyListeningError``.
+- ``TChannel.advertise()`` will now automatically start listening for connections
+  if ``listen()`` has not already been called.
+- Use ``threadloop==0.4``.
 - Removed ``print_arg``.
+
+
+0.11.2 (2015-07-20)
+-------------------
+
+- Fix sync client's advertise - needed to call listen in thread.
 
 
 0.11.1 (2015-07-17)